Deprecate JobRunner
